requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
# Remove: passlib[bcrypt]==1.7.4  # This causes the issue
passlib==1.7.4  # Without bcrypt
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0